backend/requirements.txt
fastapi
uvicorn
sqlalchemy
feedparser
spacy
pandas
rake-nltk
python-dateutil
